refactor: extract redis message history
